+++ client-wot/src/main/java/io/github/divinator/wgapi/method/wot/ClanMethod.java
@@ -1,11 +1,9 @@
 package io.github.divinator.wgapi.method.wot;
 
 import com.google.gson.reflect.TypeToken;
-import io.divinator.wgapi.client.AbstractMethodBlock;
-import io.divinator.wgapi.client.Cluster;
-import io.divinator.wgapi.client.*;
-import io.divinator.wgapi.entity.JsonResponse;
-import io.divinator.wgapi.entity.wot.clan.*;
+import io.github.divinator.wgapi.client.*;
+import io.github.divinator.wgapi.entity.JsonResponse;
+import io.github.divinator.wgapi.entity.wot.clan.*;
 
 import java.util.List;
 import java.util.Map;