+++ client-wot/src/main/java/io/github/divinator/wgapi/method/wot/GlobalMapMethod.java
@@ -1,10 +1,10 @@
 package io.github.divinator.wgapi.method.wot;
 
 import com.google.gson.reflect.TypeToken;
-import io.divinator.wgapi.client.AbstractMethodBlock;
-import io.divinator.wgapi.client.*;
-import io.divinator.wgapi.entity.JsonResponse;
-import io.divinator.wgapi.entity.wot.globalmap.*;
+import io.github.divinator.wgapi.client.AbstractMethodBlock;
+import io.github.divinator.wgapi.client.*;
+import io.github.divinator.wgapi.entity.JsonResponse;
+import io.github.divinator.wgapi.entity.wot.globalmap.*;
 
 import java.util.List;
 import java.util.Map;