+++ client-wot/src/main/java/io/github/divinator/wgapi/method/wot/TanksMethod.java
@@ -1,13 +1,13 @@
 package io.github.divinator.wgapi.method.wot;
 
 import com.google.gson.reflect.TypeToken;
-import io.divinator.wgapi.client.AbstractMethodBlock;
-import io.divinator.wgapi.client.*;
-import io.divinator.wgapi.client.Cluster;
-import io.divinator.wgapi.client.WgApiUriBuilder;
-import io.divinator.wgapi.entity.JsonResponse;
-import io.divinator.wgapi.entity.wot.tanks.Achievement;
-import io.divinator.wgapi.entity.wot.tanks.Stats;
+import io.github.divinator.wgapi.client.AbstractMethodBlock;
+import io.github.divinator.wgapi.client.*;
+import io.github.divinator.wgapi.client.Cluster;
+import io.github.divinator.wgapi.client.WgApiUriBuilder;
+import io.github.divinator.wgapi.entity.JsonResponse;
+import io.github.divinator.wgapi.entity.wot.tanks.Achievement;
+import io.github.divinator.wgapi.entity.wot.tanks.Stats;
 
 import java.util.List;
 import java.util.Map;