+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Add logging
 
 ### [2.1.0]() (2020-11-14)
+* Merge projects client and extensions
+* package name to io.github.divinator
 * Add properties file
 * Add proxy settings
 * Refectories WgApiUriBuilder class